Restructured project to python3.9 module
