+++ alicebot.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+python3 -m alice
